Remove lingering use of glog
We can't remove it as a dependency, since we need to set its verbosity
level in test/logging.go still.

We can't remove it from pkg/client/clientset/versioned/clientset.go
because it's generated.

This also leaves some references to glog in docs/telemetry.md, since
it's not clear what we should recommend.