This changes from `demo-domain.com` to `example.com`. (#1400)
We don't own `demo-domain.com`, so configuring it by default presents an abundance of problems.  `example.com` is the canonical placeholder domain, so we will configure that by default.

In the future we will very likely change this to a sane default that will automatically work with intra-cluster DNS, but until then this seems like the most reasonable default.

Fixes: https://github.com/knative/serving/issues/1365