Move the `/etc/config-logging` references into `./cmd/foo` (#1377)
* Move the `/etc/config-logging` references into `./cmd/foo`

This moves the references to the volume mounted ConfigMap for logging out of `./pkg/logging` (bad assumption for a library to make), and instead exposes a form of the library function that will do the same thing given a `logging.Config` (new, see below).

`logging.Config` is a new struct similar to `DomainConfig` and `NetworkConfig`, which can be constructed from either:
1. A `corev1.ConfigMap` (usable with `configmap.Watcher` from controllers),
2. A `map[string]string` (usable with `configmap.Load` from other components).

The Revision controller now watches for logging updates through `configmap.Watcher` and stores a `logging.Config`.

`configmap.Load` is a new utility that will reconstruct the `.Data` of a `corev1.ConfigMap` from a volume mount, so that it can share the same core logic for translation into `FooConfig` and `foo.Config` objects.

The loggers constructed for our binaries `configmap.Load` the `logger.Config` at startup.

* Make Load support the Kubernets volume layout for ConfigMap.

This updates the unit test to match the K8s ConfigMap volume layout, and support it.