Generalize the controller table test loop.
This moves the implementation of the table test the controllers are moving towards into a common location.  Since the loop is already designed to be fairly generic, this is a relatively straightforward move with no appreciable change in the logic from what was in Revision's `table_test.go`.  This does effectively backport some of the improvements made to the core logic to Service and Configuration.