Revision controller table testing. (#1437)
* Revision controller table testing.

This sets up the skeleton for writing table tests for the Revision controller, and adds a number of tests for it.  The bulk of the Revision controller behavior is now covered by the table-based test.  The notable gaps in coverage where `revision_test.go` still meaningfully increases coverage are the fluentd and VPA paths, which are flag guarded.  Covering these gaps is possible via this method, but they must be switched to `.Get()` through an informer lister that we can fake and the flag to enable this functionality passed.

Another notable difference from `revision_test.go` is that the table test doesn't make the same effort to test things like label propagation, which IMO should be done through more focused unit testing of the utilities that Reconcile builds upon instead of through Reconcile itself.

I believe this is far enough along that we should focus on adding new testing here instead of `revision_test.go` whenever possible.

Related: https://github.com/knative/serving/issues/1219

* Hoist some of the boilerplate from the main loop of the table test.

* Generalize the controller table test loop.

This moves the implementation of the table test the controllers are moving towards into a common location.  Since the loop is already designed to be fairly generic, this is a relatively straightforward move with no appreciable change in the logic from what was in Revision's `table_test.go`.  This does effectively backport some of the improvements made to the core logic to Service and Configuration.

* Comment out a useful logging line that we probably don't want on by default.