Avoid a dependency cycle. (#1485)
Having resource names alongside the resource creation created a dependency cycle in the forthcoming multi-tenant autoscaler work.

The autoscaler needs to be able to use the `names.Deployment` method to determine the names of a `Revision`'s `Deployment` resources, but (currently) the way we `MakeDeployment` (specifically the queue container) depends on `autoscaler.Config`.

Naming seemed like a relatively self-contained piece to break off to cut this cycle.  This does that for the Revision controller.