Remove GetServingNamespaceName.
This removes the `GetServingNamespaceName` function, which is currently identity (with TODO).  The idea behind this identity function (with TODO) was that we might later want to enable operators to shift resource creation into a separate namespace.  However, our usage of this function has grown inconsistent, and implementing such a change faces many steeper obstacles than bringing this function back.  So for now I'm biasing towards keeping the code lean and consistent and removing it.

The reason we use same-namespace today is that many K8s abstractions we want to integrate require this namespace colocation.  Allowing the user to specify serviceaccount, configmap references, secret references, ... only work when the Pods we push those references into operate in the same namespace as the knative/serving resources.  e.g. Currently this is our only mechanism for supplying a Revision `imagePullSecrets`.

The main piece of knowledge that I've gained since leaving the TODO is that OwnerReferences (and cascaded deletions) are scoped to same-namespace.  This happens to work cross-namespace now, but we need to stop relying on that (bug not feature), see: https://github.com/knative/serving/issues/1259.