Switch TableTest to take []runtime.Object.
This switches our `TableTest` to taking a single `[]runtime.Object` and greatly reduce the boilerplate to set up the state of the world prior to `Reconcile()`.

We still use the Listers, but we set them up by switching over the type of the objects and putting them into the appropriate ones.

Fixes: https://github.com/knative/serving/issues/1519