Move `SetDefaults()` to our types. (#1521)
* Simplify cast logic in the webhook.

* Use cmpopts.IgnoreFields to avoid comparing ServingState.

* Refactor and normalize the webhook defaulting.

This reshapes the structure of the webhook defaulting so that the 3/4 resources that perform defaulting:
1. Have a remarkably similar structure, and
1. Synthesize jsonpatches post-facto after calling a function to initialize defaults.

This is a precursor to moving the defaulting logic under pkg/apis.

* Add `SetDefaults()` to our resources.

This implements a new `Defaultable` interface on each of our resources allowing `pkg/webhook` to delegate its defaulting needs to these methods.

Fixes: https://github.com/knative/serving/issues/1518

* Initial streamlining of the way we Default and Validate.

This removes the `unmarshalFoo` methods from `pkg/webhook/{service,configuration,route}.go` by making `GenericCRD` import `v1alpha1.Validatable` and `v1alpha1.Defaultable`.

At this point these three files are virtually identical excepting that `route.go` doesn't perform defaulting currently.

The same is true for `revision.go`'s defaulting, but its validation still requires the concrete type in order to verify that the `Spec` remains unchanges across mutations.

* Consolidate all defaulting and most validation.

This removes the bulk of the type-specific logic, in favor of delegation to `v1alpha1.Defaultable` and `v1alpha1.Validatable`.