Disallow ports in network probes (#1525)
* Disallow port in network probes.

Currently we disallow specifying ports in the container spec because we require the container to listen on 8080.  This extends that to apply to probes as well, which should exclusively happen on 8080.

For HTTP probes, it is especially important that we probe exclusively on 8080 because we rewrite the port of the probe to go through our queue proxy (which then proxies to 8080 on the user container).

Fixes: https://github.com/knative/serving/issues/684

* Rework the HTTP Readiness probe port defaulting.

This extends the probe port defaulting for HTTP Readiness probes to apply to: {tcp, http} x {readiness, liveness} probes.

* Drop port from test and samples.

* improve the `name:` of tests.