Add ServiceName to status for resources that create services
For resoruces that own a core Kubernetes Services, the name of the
service should be exposed on the resoruce's status. The Revision
resources already exposes `.status.serviceName`, this PR adds the same
property for the Route and Knative Service resources.

Yes, it's a bit strange to have a Serivce include the Service name in
it's status, ¯\_(ツ)_/¯

Fixes: #1584