Make verify-codegen.sh not mess up with working tree (#1579)
Partially addresses #1575.

`verify-codegen.sh` calls `update-codegen.sh` (which in turns calls `update-deps.sh`) to autogenerate new code, so it can diff the current code in the working tree against the latest autogenerated code (and thus check if everything is up-to-date).

This might leave the working tree in a changed state. This PR undoes all changes to //pkg, //vendor and Gopkg.lock that might be caused by the `update-*.sh` scripts. This way, when running `verify-codegen.sh`, it's guaranteed that your working tree is untouched when the script finishes, as expected.

Bonus: remove redundant environment variables, and use `library.sh` for common/shared stuff, just like the other scripts.