Update global metric logger to be contextual (#1366)
Initially this was an experiment to use a global logging object, for
2 reasons:

1. The opencensus metrics exporter has no choice but to be global
2. It clutters up the function signatures a bit

After that experiment I have decided that it is not a good idea to
use a global logger because:

1. It promotes bad practices
2. What if we ever want to write tests FOR our test libraries? (if they
   become sophisticated enough) And this _may_ lead to some of these
   libs one day being actual libraries, in which case they should follow
   best practices for libraries.

So instead I have updated the metrics emitter to also be contextual, and
sadly there is no way to avoid it being global :( which means that if
tests are run in parallel within the same process, each test will be
emitting metrics using the context of whichever test had most recently
created a logger :( :( :( The only way around this I can think of is to
use a metrics emitter that is not global (i.e. a different lib).

This changes the output of metrics b/c the context will be included as
well.

A possible future modification would be to create structs for some of
these libs and initialize those structs with a logger object so the
function signatures do not include a logger.