Run conformance and E2E tests concurrently
This saves about 5 minutes of run time.

This also simplifies the running part of the E2E/conformance tests in the `e2e-tests.sh` script, as actually there's no reason to run these tests sequentially anynore.

However, in order to achieve that, the namespaces and test images location had to be unified for both set of tests (although this doesn't affect their run).