Don't ignore failures in the admission webhook (#1695)
By default, admission webhook failures are ignored. So if the webhook
service is down or can't be communicated with, the admission process
will happily ignore that and continue with the resoure changes.

By setting the FailurePolicy to Fail, we ensure that problems
communicating with the webhook service don't get silently ignored,
leading to problems later on.