Invert the relationship between controller.Base and derivatives.
The way we use `controller.Base` today is very reminiscent of OOP; it is effectively trying to be a base class, and many things are painful trying to write this way in Go (e.g. we cannot simply make `Reconcile` an abstract method to be implemented by derivatives).

With this change, we will instead move to a model where `Base` becomes `Impl` and takes an instance of the following interface:
```go
package controller

type Reconciler interface {
  Reconcile(key string) error
}
```

In this model, `controller.Impl` is the single shared controller implementation, and the `NewController` methods become responsible for instantiating it with the appropriate implementation of `Reconciler` and wiring the appropriate events up via the informers.

This cleans up the separation between `./pkg/controller` and `./pkg/controller/foo`, and will hopefully start to make the latter quite small.  Ultimately the plan is to share the former via `knative/pkg`.

Fixes: https://github.com/knative/serving/issues/1505