Enable e2e tests to pass on Minikube by tagging test images
This provides the ability to override the default 'latest' tag and
avoid its default pull policy of 'Always'. This enables the tests to
run on minikube without its internal docker daemon to be configured
with an upstream registry.

This is completely opt-in: the default behavior (using 'latest') is
unchanged.

To make the tests pass on minikube (assuming LoadBalancer type
services are routed correctly):

  eval $(minikube docker-env)
  ./test/upload-test-images.sh SOME_TAG
  go test -v -tags=e2e -count=1 ./test/e2e --tag SOME_TAG