Rework pkg/autoscaler in terms of KPA.
This mostly reworks pkg/autoscaler to operate in terms of the KPA instead of Revisions.  This is essentially the third phase outlined in #1655, but this doesn't add logic to update the KPA status because it would be too awkward with the current interface with the RevisionSynchronizer (now KPASynchronizer).  This change is big enough so call it `3a.`, I will think on the best way to achieve `3b.` as a follow up.

Progress towards: https://github.com/knative/serving/issues/1655