Rework the model for the Autoscaling controller.
This reworks the autoscaling controller as described [here
](https://github.com/knative/serving/issues/1655#issuecomment-412282318).
Essentially the goal is to make it `Reconcile()` KPAs, including any
mutations to its status.  To enable this using mode traditional controller
structuring we treat the metrics being collected in `./pkg/autoscaler` as
a virtual subresource and hook it into reconciliation in similar ways to
how we would hook up reconciliations when normal child resources change.

With this change the multi-tenant autoscaling logic will drive all
autoscaling decisions from `Reconcile()` based on the workqueue and
reflect the status into the KPA's Status conditions.