Rework pkg/autoscaler in terms of KPA. (#1830)
* Rework pkg/autoscaler in terms of KPA.

This mostly reworks pkg/autoscaler to operate in terms of the KPA instead of Revisions.  This is essentially the third phase outlined in #1655, but this doesn't add logic to update the KPA status because it would be too awkward with the current interface with the RevisionSynchronizer (now KPASynchronizer).  This change is big enough so call it `3a.`, I will think on the best way to achieve `3b.` as a follow up.

Progress towards: https://github.com/knative/serving/issues/1655

* Rework the model for the Autoscaling controller.

This reworks the autoscaling controller as described [here
](https://github.com/knative/serving/issues/1655#issuecomment-412282318).
Essentially the goal is to make it `Reconcile()` KPAs, including any
mutations to its status.  To enable this using mode traditional controller
structuring we treat the metrics being collected in `./pkg/autoscaler` as
a virtual subresource and hook it into reconciliation in similar ways to
how we would hook up reconciliations when normal child resources change.

With this change the multi-tenant autoscaling logic will drive all
autoscaling decisions from `Reconcile()` based on the workqueue and
reflect the status into the KPA's Status conditions.