Bypass istio to prevent activator retry loop.
When the activator proxies and retries requests to make sure a Revision is actually routable, Istio will pick the requests up and forward them to the activator itself if network programming is not done completely.

Resetting the Host header bypasses istio so the Revision's service can be adressed directly.

Fixes #1907