Fix e2e/conformance tests on minikube (#1840)
* Enable e2e tests to pass on Minikube by tagging test images

This provides the ability to override the default 'latest' tag and
avoid its default pull policy of 'Always'. This enables the tests to
run on minikube without its internal docker daemon to be configured
with an upstream registry.

This is completely opt-in: the default behavior (using 'latest') is
unchanged.

To make the tests pass on minikube (assuming LoadBalancer type
services are routed correctly):

  eval $(minikube docker-env)
  ./test/upload-test-images.sh SOME_TAG
  go test -v -tags=e2e -count=1 ./test/e2e --tag SOME_TAG

* Skip TestContainerErrorMsg when using a local docker repo

It's a fairly brittle test, as it's expecting a specific message
returned only by GCR, afaik

* Simplify minikube installation with minikube-lb-patch

Until minikube supports LoadBalancer type services, the patch
simplifies using knative when deployed to minikube:

 $ sudo ip route add $(cat ~/.minikube/profiles/minikube/config.json | jq -r ".KubernetesConfig.ServiceCIDR") via $(minikube ip)
 $ kubectl run minikube-lb-patch --replicas=1 --image=elsonrodriguez/minikube-lb-patch:0.1 --namespace=kube-system

* Factor image names into helper function, fixes #1707