Eliminate loggerWith... functions. (#1935)
* Update knative/pkg dependency and fix tests.

* Eliminate loggerWith... functions.

This biases towards fetching loggers from context, which now contains the reconcile key with the latest knative/pkg.  This shouldn't lose the resource sensitivity because all of our reconcilers include the resource via the controller key.

As part of this, I went a little crazy trying to root out loggerWith* functions.  In some cases this simply manifests as inlining it, which in the few remaining cases felt clearer.  This also drops our `logkey` package, biasing towards the key attribute almost everywhere it was used.