Add GetGroupVersionKind to our CRD objects (#2014)
We currently do type checking to determine an object's Kind. This is
problematic because it requires us to add each CRD version to
owner_references.go manually, which makes adding a new CRD version
tedious.

This method name is odd (non-idiomatic GetFoo()) because:
1. The embedded TypeMeta already implements GroupVersionKind(), but it
is currently empty (https://issues.k8s.io/3030), and we'd like to ensure
that our types implement this interface.
2. It mirrors the other required method, GetObjectMeta.