Report clear error on missing configuration in observed-concurrency test.
Due to a limitation of wrk's lua script, it's not possible to obtain the parameters passed to the wrk call itself inside the lua script. To work around that, the script needs an environment variable 'concurrency' set to the amount of connections configured for wrk. If that's not present, the script will panic in the end.

This adds a validation step to exit early if the variable is not present.

Fixes #2012.