Fix race condition in breaker test (#1984)
Fix race condition in `pkg/queue/breaker_test.go` which results in occasional
deadlocks and flakey tests. The order that requests were performed was not
deterministic, but the tests expect them to be ordered.

The `concurrentRequest` function will now ensure that a request is either
performed, enqueued, or rejected before returning, which will guarantee
the expected order.

Fixes #1984