Remove VPA
This removes the VPA integration from the Revision controller to simplify things.

In the KPA architecture, the VPA doesn't belong here anyways, and before it is added back I'd like e2e tests making sure it actually WAI so that changes don't break it or the code will just rot.

Fixes: https://github.com/knative/serving/issues/1876