Add '.status.imageDigest' to Revision (#2197)
* Add '.status.imageDigest' to Revision

This change moves image digest resolution from Deployment creation
to Revision reconciliation, and introduces a new field on Revision
status. This change preseves the option to list registries to skip
in which case the imageDigest field will not be populated. When the
imageDigest field is populated, it is used for Deployment creation.

Fixes #2064

* Resolve comments

* Update spec documentation and change Resolve interface

* Update deploy test and add conformance test

* Fix rebase test error and typo