Avoid emitting headers that latter reject requests
The activator set the `Knative-Activator-Num-Retries` header on requests
it forwards to a starting revsion as well as on responses returned from
the activator. When the activator recieves a new request with the
`Knative-Activator-Num-Retries` header it immediatly returns a 503. This
has the effect of causing a request that hits the activator to fail if
the request is triggered from the response that also hit the activator.

This PR makes two distinct changes to prevent this situation from
recurring:
- the request and response headers now use distinct names. The request
  header is `knative-activator-attempts`, while the response header
  is preserved as `knative-activator-num-retries`
- the request header is stripped off the request in the queue-proxy
  along side the other activator headers

Fixes #2234
Fixes knative/eventing#523