Don't keep ClusterIP in placeholder service spec (#2352)
* Don't keep ClusterIP in placeholder service spec

In #1789 we switched this to an ExternalName Service. Services created in 0.1
will still have ClusterIP set, which is Forbidden for ExternalName Services.
Ensure that we drop the ClusterIP if it is set in the spec.

* Add test for external name

Ensure that if the external name is set to something incorrect, it gets
reconciled back.