Don't rely on TypeMeta to be properly populated.
In the Route controller, we were relying on accessors on TypeMeta to tell us the
GroupVersionKind of Configurations and Revisions so that we could construct an
ObjectReference to them with slightly less boilerplate.  It turns out, however,
that the presence of TypeMeta is
[unreliable](https://github.com/kubernetes/apiextensions-apiserver/issues/29) and
that we should instead simply synthesize these directly.

Fixes: https://github.com/knative/serving/issues/2372