[RELEASE 0.2] Pull in proposed cherry-picks (#2383)
* Don't rely on TypeMeta to be properly populated. (#2373)

In the Route controller, we were relying on accessors on TypeMeta to tell us the
GroupVersionKind of Configurations and Revisions so that we could construct an
ObjectReference to them with slightly less boilerplate.  It turns out, however,
that the presence of TypeMeta is
[unreliable](https://github.com/kubernetes/apiextensions-apiserver/issues/29) and
that we should instead simply synthesize these directly.

Fixes: https://github.com/knative/serving/issues/2372

* Return error on lastPinned update failures (#2333)

As is we only log an error in this case. Fail the reconciler to ensure
we retry on failures as revisions will otherwise be GCd incorrectly.

* Guard every connection access with a non-nil check. (#2361)