Start to refactor the way builds are created.
We want to be able to elide build creation when that portion of the `ConfigurationSpec` remains unchanged.  To support those, we need to remove assumptions that `MakeRevision` can directly synthesize an `ObjectReference` for the `Build` directly from the `Configuration`.

The idea is that in a subsequent change, we will start to label builds using the schema outlined in #439 and potentially pass in an `ObjectReference` we have looked up instead.

Progress towards: #439