Fix paren and K8S_USER_OVERRIDE docs
`K8S_USER_OVERRIDE` is only needed to run the command to setup your
current user as a cluster admin, so it isn't needed as part of the
general environment setup.

Updated the DEVELOPMENT.md to explain that the values for this are
different if you're using a GKE cluster vs a minikube cluster.