Refer to knative/docs cluster setup instead of duplicating (#2241)
* Refer to knative/docs cluster setup instead of duplicating

The DEVELOPMENT.md in knative/serving referred to a doc on setting up a
kubernetes cluster (either in GKE or with minikube) which had fallen out
of date with very similar installation docs in knative/docs.

I ran into this when trying to figure out the correct scopes to use for
creating a cluster which could pass the knative/build-pipeline kaniko
integration test (https://github.com/knative/build-pipeline/issues/150)
and it turned out that the `--scopes` in the doc referenced in this
repo are different from the `--scopes` in the knative/docs repo.(I
worked around my problem my using `storage-full`, which isn't used in
either set of docs but that's a different story!)

The minikube docs that were in this repo also contained args for
specifying the location of the cluster CA certs, but I'm assuming this
is no longer needed since knative/docs doesn't have this and
https://github.com/kubernetes/minikube/issues/1647 is resolved.

* Fix paren and K8S_USER_OVERRIDE docs

`K8S_USER_OVERRIDE` is only needed to run the command to setup your
current user as a cluster admin, so it isn't needed as part of the
general environment setup.

Updated the DEVELOPMENT.md to explain that the values for this are
different if you're using a GKE cluster vs a minikube cluster.