Disable scale-to-zero during activation (#2155) (#2386)
* Disable scale-to-zero during activation (#2155)

Currently the autoscaler will start the scale-to-zero countdown
immediately after a pod has started. This causes problems for pods that
take a while to become ready, and in pathological cases where the pod
startup time is longer than the scale-to-zero threshold, the pod is
never able to serve requests.

This fix modifies the autoscaling reconciler to ignore request metrics
until atleast one pod has successfully started and has been idle for
atleast the `scale-to-zero-idle-period` (`scale-to-zero-threshold -
scale-to-zero-grace-period`).

* Add back test cases for idle/grace period in autoscaling reconciler

* Don't return errors in KPA reconcilation for control flows

* Add tests for KPA status CanMarkInactive method

* Add back godoc comments for KPA CanMarkInactive / CanScaleToZero