Don't create builds unless the BuildSpec changes. (#2402)
Things like EnvVar updates should not trigger a fresh build, when a Build is specified in the ConfigurationSpec.  This builds on prior changes to lookup a pre-existing build with the same spec before creating its own.  This modifies the Build e2e tests to each perform an env-var only update to the configuration, and check that there are no differences in the resulting BuildRef.

Fixes: https://github.com/knative/serving/issues/439