SetDefaults at the onset of Reconcile.
As part of chasing https://github.com/knative/serving/issues/639, I ran into a problem with the newly introduced `timeoutSeconds` in the upgrade test.  Basically, the Reconciler has been modified to assume that the *new* defaulting via the webhook was happening, and setting an otherwise `nil` field.  However, in the upgrade test, we reconcile objects that only went through the older webhook, so our defaulting hasn't necessarily run.

To guard against these kinds of assumptions, which may exist all over, I have added a call to `.SetDefaults` at the onset of every `Reconcile`, so that folks can simplify how they think about reconciliation.  These changes will never be propagated back to the API because we only post updated to `.Status` copying the original object for the remainder of the content.