Add tests that verify the service's status, when route and configuration are ready  (#2620)
* Add a test that verifies that pinned behaviour via Release

The PinnedType is deprecated according to the comments,
so verify the same behavior is correct, by following the setup
suggested in the PinnedType docs.

Some minor cleanup along the way.

* Add the tests to verify the status of the service after route and config are ready.

* Remove the WithPinnedRelease2 function and just use WithReleaseRollout.

* Fix code review issues.

* Update the constant name to easier differntiate the entries in the logs.

* Reconcle pinned UTs and ready state UTs

* Fix pr comments about comments.

* Update the comments to be more detailed.

Make note about which mutations are happening, and what do
we expect in the service's status to be present, when route
is ready.

* Rewrite the comments one more time.

* Merge with the eventing change, which requires to update the tests.