This change makes Revision reconcile Deployments. (#2632)
Previously we had problems with Deployment reconciliation because of the defaulting performed in the API Server.  With this change we attempt an Update if the Spec is not what we think it should be, but only treat it as a change if the Deployment we get back from the update is different from what we started from.  In this way, we avoid actually changing the deployment and triggering infinite reconciliations.

This change is sufficient for us to trigger Deployment reconciles (e.g. to replace sidecars) when folks apply a new version, but we will need https://github.com/knative/serving/issues/2332 as well for ConfigMap only changes to trigger a similar global resync.

Note that as part of this change, we need to filter out volatile annotations like `lastPinned`, which we put into the desired `PodSpec` which would cause us to pointlessly update the Deployment and trigger a rollout.

Fixes: https://github.com/knative/serving/issues/639