Conformance test coverage for secrets and configs via env var (#2751)
This changeset adds conformance test to verify propagation of secrets
and configs(via ConfigMaps) to service container as environment variables.

Test re-uses test_images/environment image to fetch environment variables from
the service. In the process runtime_conformance_helper is renamed to
conformance_helper and fetchEnvInfo() is refactored to support this
change.