Adding Default FileSystems conformance test (#2738)
This change adds a conformance test to validate Default Filesystems
requirements as specified in the Knative serving runtime-contract. The
test extends the environment test-image to expose filepath?path=<path-to-file>
API to retrieve file information for the path specified as query param.

Note: This conformance test does not test "/dev/log" requirement specified in
the runtime-contract, which will we added in the future.