Make use of the status subresource for our custom resources (#2829)
* Enable status subresource and use the status endpoint for updates

* Mark the current revision naming scheme as deprecated

* Label revisions with the configuration's metadata generation

* When reconciling a config use metadata.generation to find the latest revision

If a revision cannot be found it will fall back and try to query it
through the revision name (which uses spec.generation).

After a config is reconciled a latest created revision should have
a new label specifying the configuration's metadata generation.

* status.observedGeneration now observes metadata.generation

This change applies to Configuration and Service types

* include addition assertions to prevent bad expectations and extra subresource updates