Fix race condition in TestScaleToX. (#2883)
The functions that alter the state of a test (Error, Fatal and their formatting counterparts mainly) are not to be called in a goroutine which is not the goroutine running the test itself. It leads to a race condition.

This fixes this race condition and propagates any errors to the main control loop of the test to be handled accordingly.