TCP probe the user-container from the queue-proxy before marking the pod as ready.
As we cannot easily run the TCP probe from kubernetes to check the user-container is up because of Istio interfering, this adds a TCP probe that's run as part of the queue-proxy's readinessProbe. The queue-proxy will only report itself ready once the TCP probe to the user-container has succeeded.

This also externalizes the health state as kept in the queue-proxy to enable testability of that code.