Clean up logging in controller main files (#2913)
* Clean up logging in controller main files

There is a lot of inconsistent use of the Zap logging functionality
throughout the project. This change moves all of the controller main
functions to use the Zap With() syntax when errors are present. This
fixes a bug in output for logging calls that forgot the "w" at the end of
the logging output level, and changes sprintf formatted error messages
to the use context options of Zap.

This change targets just these files to limit conflicts generated by a
larger single commit.

Previous log line without "w" or "f":
{"level":"info","msg":"Previous message.{error 25 0  Example Error}"}
Previous log line with "f":
{"level":"info","msg":"Previous message: Example Error"}
New log line with "w":
{"level":"info","msg":"New Message","error":"Example Error"}

* Update cmd/autoscaler/main.go

* Update cmd/controller/main.go

* Update test/controller/main.go