TCP probe the user-container from the queue-proxy before marking the pod ready. (#2915)
* TCP probe the user-container from the queue-proxy before marking the pod as ready.

As we cannot easily run the TCP probe from kubernetes to check the user-container is up because of Istio interfering, this adds a TCP probe that's run as part of the queue-proxy's readinessProbe. The queue-proxy will only report itself ready once the TCP probe to the user-container has succeeded.

This also externalizes the health state as kept in the queue-proxy to enable testability of that code.

* Fix some variable shadowing.

* Inline and shorten prober function.

* Add slash to constants.

* Fix logging and remove unnecessary variable.

* Remove redundant status code.

* Use switch over multiple ifs.

* It's 2019.

* Remove stutter.

* Tune back logging.

* Unify userTargetAddress.

* Add some more logging touches.

* Finish testing.

* Drop colon.

* Lameduck only when shutting down.

* Report the last error returned by the poller.

* Mark helper methods, rename to 'want'.

* Localize assert methods.

* Fix error logging.

* Make test more meaningful.