Add a readinessProbe and timeout to the scale test. (#2904)
The timeout is useful because otherwise we use a termination grace period matching the RevisionTimeout, which defaults to 5 minutes and feels like it takes forever when you have 50+ Services.

The readinessProbe is to ensure that we don't send traffic to the Service expecting it to pass before it's actually ready.

Related to: #2850