Reformulate the e2e "probing" in terms of SLI/SLO.
This reformulates the prober logic we use (heavily) during the scale test to operate in terms of SLI (measured success) and SLO (target success) at two levels:
1. Per-probe, which could potentially drift lower.
1. Aggregate, which should be quite high.

As part of this refactoring, I uncovered the source of our woes with this test, which is captured in detail at #2946.  The basic idea is that we disable autoscaling (in particular activation) because it is unreliable at the level of churn that these scale tests create with high values of N.

With N=200, and no mesh (simply what I run day-to-day) I was able to pass 3x in a row with an Aggregate SLI of 100% (no dropped probes).  This is night/day from what we were seeing before.

Fixes: https://github.com/knative/serving/issues/2850

I still need to explore (but out of scope for the above issue):
1. Whether things fail at 200 with the mesh
1. How high we can go with/without the mesh.
1. Raise the high-watermark that we test to be beyond 50