Release mode processing for the route not being ready. (#2948)
* add interface

* Implement release mode checking for the service not yet being ready.

When we update the service in the release mode, but the route is not yet
updated to the new version, we should mark the service to be in `Unknown` ready state.
See: #2430

* remove debug logging

* Bring back the accidentaly deleted comment.

* Apply suggestions from the code review.

* Replace cmp with kmp package.

* Update pkg/reconciler/v1alpha1/service/service.go

* Remove the check for type of service

Since reconcile() is only invoked, when we're not in manual mode,
hence the type check is useless here.
Also move the DeepCopy() one word right, saving us a few microseconds.

* diff -> eq