Switch to using Finalizers for ClusterIngress deletion.
This removes the OwnerReferences from ClusterIngress, which are buggy and unsupported anyways.

Instead, we will add a finalizer prior to creating ClusterIngress, which will enable us to hook into Route deletion to clean up the ClusterIngress resources manually.

When we see a Route with a DeletionTimestamp, we elide most processing.  When we are the first finalizer on the list, it is our turn and we delete the cluster ingresses matching our label selector and remove the finalizer to allow deletion to proceed.

Fixes: https://github.com/knative/serving/issues/2570