Pull in the knative/pkg update to make Severity less scary. (#3038)
With this change, `Severity: Error` is always written to the API Server as `Severity: ""`.

This is a backwards compatible change because for (drumroll) backwards compatibility in clients, we elected to make `"Error"` the default when omitted (so that the previously 100% terminal world w/o `Severity` mapped cleanly onto the new world).

Fixes: https://github.com/knative/serving/issues/2507