HTTP2 and gRPC support (#2539)
* Use http.DefaultTransport dialer settings in h2c.DefaultTransport

* Support streaming in activator/util.Rewinder

* Use custom timeout handler in queue-proxy

The http.TimeoutHandler will buffer the response body in memory
until either the request completes or the request times out.

This works well for HTTP, but is a problem for HTTP2 and gRPC
streaming requests, where responses should be written as each
sub-request is processed.

This commit enforces the timeout by processing the request in a separate
goroutine and panicing with http.ErrAbortHandler if the timeout is
reached. The http.ErrAborthandler is a canary error used by the net/http
and x/net/http2 packages to gracefully end the connection without dumping
the stack trace.

* Use http2 instead of http as the k8s service port names

Istio uses port names in k8s services to determine protocols supported
by the service.

This change allows kservices to support HTTP/2 and gRPC traffic.

* Enforce max content length for streaming requests in activator

* Don't read original reader again after rewinder is closed

* Use chan struct{} instead of chan bool in queue-proxy timeoutHandler

* Move LimitReadCloser from activator handler to activator util and add test coverage

* Add/fix godoc comments for activator/util

* Explictly set Dialer config defaults in h2c.DefaultTransport

* Remove timeoutHandler from queue-proxy

* Revert k8s Service port name from http2 to http

Changing it to http2 for all services breaks services which only support
http1. Support for http2 will require selectively setting the port name
to http2 only for services that explicitly support it.

* Run activator on two ports, to support activating both http1 and h2c targets

* Add RevisionProtocolType to API

* Dynamically select 'http' or 'http2' for k8s service port name

* Dynamically route to the http1 or h2c activator service port based on the revision protocol

* Add test coverage for activator.ServicePort

* Print errors in EnforceLengthHandler tests